refactor(flutter): extract platform abstraction for web/native split
Move all platform-specific logic behind a QuilttPlatformInterface.
Mobile implementation extracted to quiltt_connector_mobile.dart.
Telemetry split into native/web stubs so the public telemetry_utils.dart
compiles on web without importing dart:io.
