+++ packages/flutter/lib/telemetry_utils.dart
@@ -1,40 +1,11 @@
-import 'dart:io' show Platform;
-
-import 'package:flutter/foundation.dart';
+// telemetry_native.dart uses dart:io which doesn't exist on web. The conditional
+// export swaps in the web stub so that this public library file compiles on all
+// supported platforms, even though getRuntimePlatformInfo() is only *called* by
+// the mobile path (quiltt_connector_mobile.dart).
+export 'src/platform/telemetry_native.dart'
+    if (dart.library.js_interop) 'src/platform/telemetry_web.dart';
 
 /// Returns the SDK User-Agent string in the format `Quiltt/<version> (<platform>)`.
 String getSDKAgent(String sdkVersion, String platformInfo) {
   return 'Quiltt/$sdkVersion ($platformInfo)';
-}
-
-/// Returns a human-readable string describing the current runtime platform and OS version.
-String getRuntimePlatformInfo() {
-  if (kIsWeb) {
-    return 'Flutter Web';
-  }
-
-  final targetPlatform = switch (defaultTargetPlatform) {
-    TargetPlatform.android => 'Flutter Android',
-    TargetPlatform.iOS => 'Flutter iOS',
-    TargetPlatform.macOS => 'Flutter macOS',
-    TargetPlatform.windows => 'Flutter Windows',
-    TargetPlatform.linux => 'Flutter Linux',
-    TargetPlatform.fuchsia => 'Flutter Fuchsia',
-  };
-
-  final operatingSystemVersion = _normalizeWhitespace(
-    Platform.operatingSystemVersion,
-  );
-  final dartVersion = _extractDartVersion(Platform.version);
-
-  return '$targetPlatform; OS $operatingSystemVersion; Dart/$dartVersion';
-}
-
-String _extractDartVersion(String version) {
-  final firstToken = version.trim().split(RegExp(r'\s+')).first;
-  return firstToken.isEmpty ? 'Unknown' : firstToken;
-}
-
-String _normalizeWhitespace(String value) {
-  return value.replaceAll(RegExp(r'\s+'), ' ').trim();
 }
